compose bom 버전 업데이트, 온도 자료형 Short로 변경 중

+++ core/network/src/main/java/io/github/pknujsp/everyweather/core/network/api/kma/KmaResponse.kt
@@ -15,8 +15,8 @@ class KmaCurrentWeatherResponse(
 ) : CurrentWeatherResponseModel {
 
     val weatherCondition: String = hourlyForecast.weatherCondition
-    val temperature: Double = currentWeather.temperature
-    val feelsLikeTemperature: Double = currentWeather.feelsLikeTemperature
+    val temperature: Short = currentWeather.temperature
+    val feelsLikeTemperature: Short = currentWeather.feelsLikeTemperature
     val humidity: Int = currentWeather.humidity
     val windSpeed: Double = currentWeather.windSpeed
     val windDirection: Int = currentWeather.windDirection
