+++ core/network/src/main/java/io/github/pknujsp/everyweather/core/network/api/kma/parser/ParsedKmaCurrentWeather.kt
@@ -3,9 +3,9 @@ package io.github.pknujsp.everyweather.core.network.api.kma.parser
 
 data class ParsedKmaCurrentWeather(
     val dateTime: String,
-    val temperature: Double,
-    val yesterdayTemperature: Double,
-    val feelsLikeTemperature: Double,
+    val temperature: Short,
+    val yesterdayTemperature: Short,
+    val feelsLikeTemperature: Short,
     val humidity: Int,
     val windDirection: Int,
     val windSpeed: Double,
